+++ src/feature/common/Header/Header.tsx
@@ -1,0 +1,14 @@
+/**
+ * Header
+ * * 共通Titleを持つ
+ * * 共通Navigationを持つ
+ */
+import { Navigation } from './Navigation';
+
+export const Header = () => {
+  return (
+    <>
+      <Navigation />
+    </>
+  );
+};
