+++ src/feature/common/README.md
@@ -1,0 +1,4 @@
+# common ディレクトリの役割
+
+- app の layout で利用する feature を格納する
+- src/feature の各 Components で利用する UI を格納する
